+++ polaris-core/src/main/java/org/apache/polaris/core/policy/TransactionalPolicyMappingPersistence.java
@@ -95,7 +95,7 @@ public interface TransactionalPolicyMappingPersistence {
       @Nonnull PolarisCallContext callCtx,
       long policyCatalogId,
       long policyId,
-      long policyTypeCode) {
+      int policyTypeCode) {
     throw new UnsupportedOperationException("Not Implemented");
   }
 }
